Minor: rename glomap cost_function.h to cost_functions.h.
