+++ src/glomap/estimators/cost_functions.h
