+++ src/glomap/estimators/cost_functions_test.cc
@@ -27,7 +27,7 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#include "glomap/estimators/cost_function.h"
+#include "glomap/estimators/cost_functions.h"
 
 #include "colmap/util/eigen_matchers.h"
