+++ src/glomap/estimators/view_graph_calibration.cc
@@ -2,7 +2,7 @@
 
 #include "colmap/scene/two_view_geometry.h"
 
-#include "glomap/estimators/cost_function.h"
+#include "glomap/estimators/cost_functions.h"
 
 #include <thread>
